semisolved morsecode led, more methods to do

+++ MorseCodeER/Timer.cpp
@@ -4,21 +4,23 @@ unsigned long Timer::reset()
 {
   return programStart = millis();
 }
-bool Timer::millisCounter(int wantedMilliseconds)
+
+bool Timer::millisCounter(float wantedMilliseconds)
 {
   return (millis() - programStart >= oneMillisecond * wantedMilliseconds);
 }
-bool Timer::secCounter(int wantedSeconds)
+
+bool Timer::secCounter(float wantedSeconds)
 {
   return (millis() - programStart >= oneSecond * wantedSeconds);
 }
 
-bool Timer::minCounter(int wantedMinutes)
+bool Timer::minCounter(float wantedMinutes)
 {
   return (millis() -programStart >= wantedMinutes * oneMinute);
 }
 
-bool Timer::hourCounter(int wantedHours)
+bool Timer::hourCounter(float wantedHours)
 {
   return (millis() - programStart >= wantedHours * oneHour);
 }
